test(bundle_tag_parser): move unit tests
