+++ test/bundle_tag_parser_test.html
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+    <title>Deserialize Tags Test</title>
+    <script src="js/avsc.min.js"></script>
+    <script src="js/tagparser.js"></script>
+    <link rel="x-dart-test" href="bundle_tag_parser_test.dart" />
+    <script defer src="packages/test/dart.js"></script>
+</head>
+
+<body>
+    // ...
+</body>
+
+</html>
